1231 - enemy 킬카운트 골드 정수화 보스 부모클래스 연결

+++ Source/MegaBonkClone/Private/UI/TopDetailWidget.cpp
@@ -111,7 +111,7 @@ void UTopDetailWidget::RefreshTimerText(float InMin, float InSec)
 
 void UTopDetailWidget::RefreshGoldCoinText(float InGold)
 {
-	FText goldCoinText = FText::AsNumber(InGold);
+	FText goldCoinText = FText::AsNumber(static_cast<int32>(InGold));
 
 	GoldCoinText->SetText(goldCoinText);
 }
